compiler: merge CIL-related assemblies (2)
