+++ src/compiler/backend/cil/CilGenerator.Compile.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using IKVM.Reflection;
 using IKVM.Reflection.Emit;
+using Uno.Compiler.API.Backends;
 using Uno.Compiler.API.Domain.IL;
 using Uno.Compiler.API.Domain.IL.Expressions;
 using Uno.Compiler.API.Domain.IL.Members;
